requirements.txt
discord.py>=2.7.1
Pillow>=10.0.0
wavelink>=3.4.1
python-dotenv>=1.0.0
motor>=3.3.0
pymongo>=4.6.0
aiohttp>=3.9.0
dnspython>=2.3.0
PyNaCl>=1.5.0
psutil>=5.9.0
spotipy>=2.23.0
jishaku>=2.5.0
